refactor(service): move WeatherApiService to the service package
