+++ src/main/java/pet/project/servlet/HomeServlet.java
@@ -12,7 +12,6 @@ import org.thymeleaf.ITemplateEngine;
 import org.thymeleaf.context.WebContext;
 import org.thymeleaf.web.servlet.IServletWebExchange;
 import org.thymeleaf.web.servlet.JakartaServletWebApplication;
-import pet.project.WeatherApiService;
 import pet.project.dao.LocationDao;
 import pet.project.dao.SessionDao;
 import pet.project.model.Location;
@@ -20,6 +19,7 @@ import pet.project.model.Session;
 import pet.project.model.User;
 import pet.project.model.api.WeatherApiModel;
 import pet.project.model.dto.WeatherDto;
+import pet.project.service.WeatherApiService;
 import pet.project.util.TemplateEngineUtil;
 
 import java.io.IOException;
